+++ app/src/main/java/com/infinity/omos/ui/onboarding/login/LoginViewModel.kt
@@ -6,11 +6,11 @@ import com.infinity.omos.data.user.UserCredential
 import com.infinity.omos.data.user.UserSnsCredential
 import com.infinity.omos.repository.UserRepository
 import com.infinity.omos.ui.onboarding.ErrorField
-import com.infinity.omos.ui.onboarding.OnboardingState
-import com.infinity.omos.ui.onboarding.OnboardingState.Failure.Companion.BLANK_EMAIL_ERROR_MESSAGE
-import com.infinity.omos.ui.onboarding.OnboardingState.Failure.Companion.BLANK_PASSWORD_ERROR_MESSAGE
-import com.infinity.omos.ui.onboarding.OnboardingState.Failure.Companion.INCORRECT_CONTENTS_ERROR_MESSAGE
-import com.infinity.omos.ui.onboarding.OnboardingState.Failure.Companion.NOT_EXIST_USER_ERROR_MESSAGE
+import com.infinity.omos.ui.onboarding.base.OnboardingState
+import com.infinity.omos.ui.onboarding.base.OnboardingState.Failure.Companion.BLANK_EMAIL_ERROR_MESSAGE
+import com.infinity.omos.ui.onboarding.base.OnboardingState.Failure.Companion.BLANK_PASSWORD_ERROR_MESSAGE
+import com.infinity.omos.ui.onboarding.base.OnboardingState.Failure.Companion.INCORRECT_CONTENTS_ERROR_MESSAGE
+import com.infinity.omos.ui.onboarding.base.OnboardingState.Failure.Companion.NOT_EXIST_USER_ERROR_MESSAGE
 import com.infinity.omos.utils.Pattern.Companion.emailPattern
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
