+++ app/src/main/java/com/infinity/omos/ui/setting/change/password/ChangePasswordFragment.kt
@@ -11,7 +11,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.infinity.omos.R
 import com.infinity.omos.databinding.FragmentChangePasswordBinding
-import com.infinity.omos.ui.onboarding.OnboardingState
+import com.infinity.omos.ui.onboarding.base.OnboardingState
 import com.infinity.omos.utils.repeatOnStarted
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -42,9 +42,16 @@ class ChangePasswordFragment : Fragment() {
     }
 
     private fun initListener() {
+        initBackButtonListener()
         initNewPasswordListener()
         initConfirmNewPasswordListener()
         initCompleteListener()
+    }
+
+    private fun initBackButtonListener() {
+        binding.toolbar.setNavigationOnClickListener {
+            findNavController().popBackStack()
+        }
     }
 
     private fun initNewPasswordListener() = with(binding.ofvNewPassword) {
