+++ app/src/main/java/com/infinity/omos/ui/setting/change/password/ChangePasswordViewModel.kt
@@ -6,15 +6,13 @@ import com.infinity.omos.data.user.UserId
 import com.infinity.omos.data.user.UserPassword
 import com.infinity.omos.repository.UserRepository
 import com.infinity.omos.ui.onboarding.ErrorField
-import com.infinity.omos.ui.onboarding.OnboardingState
-import com.infinity.omos.ui.onboarding.OnboardingState.Failure.Companion.NETWORK_ERROR_MESSAGE
+import com.infinity.omos.ui.onboarding.base.OnboardingState
+import com.infinity.omos.ui.onboarding.base.OnboardingState.Failure.Companion.NETWORK_ERROR_MESSAGE
 import com.infinity.omos.utils.Pattern.Companion.passwordPattern
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 @HiltViewModel
