+++ app/src/main/res/layout/fragment_change_password.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-        <import type="com.infinity.omos.ui.onboarding.OnboardingState" />
+        <import type="com.infinity.omos.ui.onboarding.base.OnboardingState" />
 
         <variable
             name="vm"
