+++ app/src/main/res/layout/fragment_forgot_password.xml
@@ -5,9 +5,9 @@
 
     <data>
 
-        <import type="com.infinity.omos.ui.onboarding.OnboardingState" />
+        <import type="com.infinity.omos.ui.onboarding.base.OnboardingState" />
 
-        <import type="com.infinity.omos.ui.onboarding.AuthCodeState" />
+        <import type="com.infinity.omos.ui.onboarding.base.AuthCodeState" />
 
         <variable
             name="vm"
