+++ app/src/main/res/layout/fragment_sign_up.xml
@@ -5,7 +5,7 @@
 
     <data>
 
-        <import type="com.infinity.omos.ui.onboarding.OnboardingState" />
+        <import type="com.infinity.omos.ui.onboarding.base.OnboardingState" />
 
         <variable
             name="vm"
